users/requirements.txt
asgiref==3.3.1
Command==0.1.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
PyJWT==2.0.1
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
